Version submitted for comment
Changed getRandomColor function to return a color object rather than just a random number.
Deleted the type check on input parameters for getRandomColor because these are supplied by the application not by a user.
updated  printQuote to use a single call to getRandomColor to retrun the new color object.
Added comments to README